+++ crates/header-translator/src/stmt.rs
@@ -140,6 +140,29 @@ fn parse_class_generics(entity: &Entity<'_>, _context: &Context<'_>) -> Vec<Stri
     generics
 }
 
+fn parse_attributes(entity: &Entity<'_>, context: &Context<'_>) -> (Option<bool>, bool) {
+    let mut sendable = None;
+    let mut mainthreadonly = false;
+
+    immediate_children(entity, |entity, _span| {
+        if let EntityKind::UnexposedAttr = entity.get_kind() {
+            if let Some(attr) = UnexposedAttr::parse(&entity, context) {
+                match attr {
+                    UnexposedAttr::Sendable => sendable = Some(true),
+                    UnexposedAttr::NonSendable => sendable = Some(false),
+                    UnexposedAttr::UIActor => {
+                        sendable = Some(false);
+                        mainthreadonly = true;
+                    }
+                    attr => error!(?attr, "unknown attribute"),
+                }
+            }
+        }
+    });
+
+    (sendable, mainthreadonly)
+}
+
 fn parse_methods(
     entity: &Entity<'_>,
     get_data: impl Fn(&str) -> MethodData,
@@ -212,7 +235,7 @@ fn parse_methods(
 /// - `EntityKind::ObjCInterfaceDecl`
 /// - `EntityKind::ObjCProtocolDecl`
 /// - `EntityKind::ObjCCategoryDecl`
-fn verify_objc_decl(entity: &Entity<'_>, context: &Context<'_>) {
+fn verify_objc_decl(entity: &Entity<'_>, _context: &Context<'_>) {
     let parent_kind = entity.get_kind();
 
     immediate_children(entity, |entity, _span| {
@@ -268,9 +291,7 @@ fn verify_objc_decl(entity: &Entity<'_>, context: &Context<'_>) {
                 // Maybe useful for knowing when to implement `Error` for the type
             }
             (EntityKind::UnexposedAttr, _) => {
-                if let Some(attr) = UnexposedAttr::parse(&entity, context) {
-                    error!(?attr, "unknown attribute");
-                }
+                // Parsed in parse_attributes
             }
             (_, parent_kind) => error!(?parent_kind, "unknown in parent"),
         }
@@ -285,7 +306,7 @@ pub enum Mutability {
     MutableWithImmutableSuperclass(ItemIdentifier),
     #[default]
     InteriorMutable,
-    // MainThreadOnly,
+    MainThreadOnly,
 }
 
 impl Mutability {
@@ -309,6 +330,7 @@ impl fmt::Display for Mutability {
                 write!(f, "MutableWithImmutableSuperclass<{}>", superclass.path())
             }
             Self::InteriorMutable => write!(f, "InteriorMutable"),
+            Self::MainThreadOnly => write!(f, "MainThreadOnly"),
         }
     }
 }
@@ -327,6 +349,7 @@ pub enum Stmt {
         derives: Derives,
         mutability: Mutability,
         skipped: bool,
+        sendable: bool,
     },
     /// @interface class_name (name) <protocols*>
     /// ->
@@ -349,6 +372,8 @@ pub enum Stmt {
         availability: Availability,
         protocols: BTreeSet<ItemIdentifier>,
         methods: Vec<Method>,
+        required_sendable: bool,
+        required_mainthreadonly: bool,
     },
     /// @interface ty: _ <protocols*>
     /// @interface ty (_) <protocols*>
@@ -377,6 +402,7 @@ pub enum Stmt {
         availability: Availability,
         boxable: bool,
         fields: Vec<(String, Ty)>,
+        sendable: Option<bool>,
     },
     /// typedef NS_OPTIONS(type, name) {
     ///     variants*
@@ -399,6 +425,7 @@ pub enum Stmt {
         ty: Ty,
         kind: Option<UnexposedAttr>,
         variants: Vec<(String, Availability, Expr)>,
+        sendable: Option<bool>,
     },
     /// static const ty name = expr;
     /// extern const ty name;
@@ -431,14 +458,22 @@ pub enum Stmt {
     },
 }
 
-fn parse_struct(entity: &Entity<'_>, context: &Context<'_>) -> (bool, Vec<(String, Ty)>) {
+fn parse_struct(
+    entity: &Entity<'_>,
+    context: &Context<'_>,
+) -> (bool, Vec<(String, Ty)>, Option<bool>) {
     let mut boxable = false;
     let mut fields = Vec::new();
+    let mut sendable = None;
 
     immediate_children(entity, |entity, span| match entity.get_kind() {
         EntityKind::UnexposedAttr => {
             if let Some(attr) = UnexposedAttr::parse(&entity, context) {
-                error!(?attr, "unknown attribute");
+                match attr {
+                    UnexposedAttr::Sendable => sendable = Some(true),
+                    UnexposedAttr::NonSendable => sendable = Some(false),
+                    attr => error!(?attr, "unknown attribute"),
+                }
             }
         }
         EntityKind::FieldDecl => {
@@ -462,7 +497,7 @@ fn parse_struct(entity: &Entity<'_>, context: &Context<'_>) -> (bool, Vec<(Strin
         _ => error!("unknown"),
     });
 
-    (boxable, fields)
+    (boxable, fields, sendable)
 }
 
 fn parse_fn_param_children(entity: &Entity<'_>, context: &Context<'_>) {
@@ -530,6 +565,8 @@ impl Stmt {
                     context,
                 );
 
+                let (sendable, mut mainthreadonly) = parse_attributes(entity, context);
+
                 let mut protocols = Default::default();
                 parse_protocols(entity, &mut protocols, context);
 
@@ -542,7 +579,17 @@ impl Stmt {
 
                 let superclasses: Vec<_> = superclasses_full
                     .iter()
-                    .map(|(id, generics, _)| (id.clone(), generics.clone()))
+                    .map(|(id, generics, entity)| {
+                        // Ignore sendability on superclasses; because it's an auto trait, it's propagated to subclasses anyhow!
+                        let (_sendable, superclass_mainthreadonly) =
+                            parse_attributes(entity, context);
+
+                        if superclass_mainthreadonly {
+                            mainthreadonly = true;
+                        }
+
+                        (id.clone(), generics.clone())
+                    })
                     .collect();
 
                 // Used for duplicate checking (sometimes the subclass
@@ -608,8 +655,13 @@ impl Stmt {
                     superclasses,
                     designated_initializers,
                     derives: data.map(|data| data.derives.clone()).unwrap_or_default(),
-                    mutability: data.map(|data| data.mutability.clone()).unwrap_or_default(),
+                    mutability: if mainthreadonly {
+                        Mutability::MainThreadOnly
+                    } else {
+                        data.map(|data| data.mutability.clone()).unwrap_or_default()
+                    },
                     skipped: data.map(|data| data.definition_skipped).unwrap_or_default(),
+                    sendable: sendable.unwrap_or(false),
                 })
                 .chain(protocols.into_iter().map(|protocol| Self::ProtocolImpl {
                     cls: id.clone(),
@@ -652,6 +704,14 @@ impl Stmt {
                     context,
                 );
 
+                let (sendable, mainthreadonly) = parse_attributes(entity, context);
+                if let Some(sendable) = sendable {
+                    error!(?sendable, "sendable on category");
+                }
+                if mainthreadonly {
+                    error!("@UIActor on category");
+                }
+
                 if !designated_initializers.is_empty() {
                     warn!(
                         ?designated_initializers,
@@ -661,7 +721,18 @@ impl Stmt {
 
                 let superclasses: Vec<_> = parse_superclasses(entity, context)
                     .into_iter()
-                    .map(|(id, generics, _)| (id, generics))
+                    .map(|(id, generics, entity)| {
+                        let (sendable, mainthreadonly) = parse_attributes(&entity, context);
+
+                        if let Some(sendable) = sendable {
+                            error!(?sendable, "sendable on category superclass");
+                        }
+                        if mainthreadonly {
+                            error!("@UIActor on category superclass");
+                        }
+
+                        (id, generics)
+                    })
                     .collect();
 
                 let subclass_methods = if let Mutability::ImmutableWithMutableSubclass(subclass) =
@@ -748,6 +819,8 @@ impl Stmt {
                     context,
                 );
 
+                let (sendable, mainthreadonly) = parse_attributes(entity, context);
+
                 if !designated_initializers.is_empty() {
                     warn!(
                         ?designated_initializers,
@@ -760,6 +833,8 @@ impl Stmt {
                     availability,
                     protocols,
                     methods,
+                    required_sendable: sendable.unwrap_or(false),
+                    required_mainthreadonly: mainthreadonly,
                 }]
             }
             EntityKind::TypedefDecl => {
@@ -776,7 +851,11 @@ impl Stmt {
                             if kind.is_some() {
                                 panic!("got multiple unexposed attributes {kind:?}, {attr:?}");
                             }
-                            kind = Some(attr);
+                            match attr {
+                                // TODO
+                                UnexposedAttr::Sendable => warn!("sendable on typedef"),
+                                _ => kind = Some(attr),
+                            }
                         }
                     }
                     EntityKind::StructDecl => {
@@ -812,7 +891,7 @@ impl Stmt {
                     _ => error!("unknown"),
                 });
 
-                if let Some((boxable, fields)) = struct_ {
+                if let Some((boxable, fields, sendable)) = struct_ {
                     assert_eq!(kind, None, "should not have parsed a kind");
                     return vec![Self::StructDecl {
                         id,
@@ -820,6 +899,7 @@ impl Stmt {
                         availability,
                         boxable,
                         fields,
+                        sendable,
                     }];
                 }
 
@@ -870,13 +950,14 @@ impl Stmt {
                     }
 
                     if !id.name.starts_with('_') {
-                        let (boxable, fields) = parse_struct(entity, context);
+                        let (boxable, fields, sendable) = parse_struct(entity, context);
                         return vec![Self::StructDecl {
                             id,
                             encoding_name: None,
                             availability,
                             boxable,
                             fields,
+                            sendable,
                         }];
                     }
                 }
@@ -907,6 +988,7 @@ impl Stmt {
                 let ty = Ty::parse_enum(ty, context);
                 let mut kind = None;
                 let mut variants = Vec::new();
+                let mut sendable = None;
 
                 immediate_children(entity, |entity, _span| match entity.get_kind() {
                     EntityKind::EnumConstantDecl => {
@@ -941,10 +1023,19 @@ impl Stmt {
                     }
                     EntityKind::UnexposedAttr => {
                         if let Some(attr) = UnexposedAttr::parse(&entity, context) {
-                            if let Some(kind) = &kind {
-                                assert_eq!(kind, &attr, "got differing enum kinds in {id:?}");
-                            } else {
-                                kind = Some(attr);
+                            match attr {
+                                UnexposedAttr::Sendable => sendable = Some(true),
+                                UnexposedAttr::NonSendable => sendable = Some(false),
+                                attr => {
+                                    if let Some(kind) = &kind {
+                                        assert_eq!(
+                                            kind, &attr,
+                                            "got differing enum kinds in {id:?}"
+                                        );
+                                    } else {
+                                        kind = Some(attr);
+                                    }
+                                }
                             }
                         }
                     }
@@ -972,6 +1063,7 @@ impl Stmt {
                     ty,
                     kind,
                     variants,
+                    sendable,
                 }]
             }
             EntityKind::VarDecl => {
@@ -1247,6 +1339,7 @@ impl fmt::Display for Stmt {
                 derives,
                 mutability,
                 skipped,
+                sendable,
             } => {
                 if *skipped {
                     return Ok(());
@@ -1263,7 +1356,8 @@ impl fmt::Display for Stmt {
                     Mutability::Immutable
                     | Mutability::Mutable
                     | Mutability::ImmutableWithMutableSubclass(_)
-                    | Mutability::InteriorMutable => id.feature(),
+                    | Mutability::InteriorMutable
+                    | Mutability::MainThreadOnly => id.feature(),
                 };
 
                 let (superclass, superclasses_rest) = superclasses.split_at(1);
@@ -1355,6 +1449,20 @@ impl fmt::Display for Stmt {
                 }
                 writeln!(f, "    }}")?;
                 writeln!(f, ");")?;
+
+                if *sendable && generics.is_empty() {
+                    writeln!(f)?;
+                    if let Some(feature) = &main_feature_gate {
+                        writeln!(f, "    #[cfg(feature = \"{feature}\")]")?;
+                    }
+                    writeln!(f, "unsafe impl Send for {} {{}}", id.name)?;
+
+                    writeln!(f)?;
+                    if let Some(feature) = &main_feature_gate {
+                        writeln!(f, "    #[cfg(feature = \"{feature}\")]")?;
+                    }
+                    writeln!(f, "unsafe impl Sync for {} {{}}", id.name)?;
+                }
             }
             Self::Methods {
                 cls,
@@ -1511,6 +1619,8 @@ impl fmt::Display for Stmt {
                 availability,
                 protocols,
                 methods,
+                required_sendable: _,
+                required_mainthreadonly: _,
             } => {
                 writeln!(f, "extern_protocol!(")?;
                 write!(f, "{availability}")?;
@@ -1532,6 +1642,15 @@ impl fmt::Display for Stmt {
                         write!(f, "{}", protocol.path())?;
                     }
                 }
+                // TODO
+                // if *required_sendable {
+                //     if protocols.is_empty() {
+                //         write!(f, ": ")?;
+                //     } else {
+                //         write!(f, "+ ")?;
+                //     }
+                //     write!(f, "Send + Sync")?;
+                // }
                 writeln!(f, " {{")?;
 
                 for method in methods {
@@ -1573,6 +1692,7 @@ impl fmt::Display for Stmt {
                 availability,
                 boxable: _,
                 fields,
+                sendable,
             } => {
                 writeln!(f, "extern_struct!(")?;
                 if let Some(encoding_name) = encoding_name {
@@ -1590,6 +1710,14 @@ impl fmt::Display for Stmt {
                 }
                 writeln!(f, "    }}")?;
                 writeln!(f, ");")?;
+
+                if let Some(true) = sendable {
+                    writeln!(f)?;
+                    writeln!(f, "unsafe impl Send for {} {{}}", id.name)?;
+
+                    writeln!(f)?;
+                    writeln!(f, "unsafe impl Sync for {} {{}}", id.name)?;
+                }
             }
             Self::EnumDecl {
                 id,
@@ -1597,6 +1725,7 @@ impl fmt::Display for Stmt {
                 ty,
                 kind,
                 variants,
+                sendable,
             } => {
                 let macro_name = match kind {
                     None => "extern_enum",
@@ -1620,6 +1749,16 @@ impl fmt::Display for Stmt {
                 }
                 writeln!(f, "    }}")?;
                 writeln!(f, ");")?;
+
+                if let Some(true) = sendable {
+                    if let Some(name) = &id.name {
+                        writeln!(f)?;
+                        writeln!(f, "unsafe impl Send for {name} {{}}")?;
+
+                        writeln!(f)?;
+                        writeln!(f, "unsafe impl Sync for {name} {{}}")?;
+                    }
+                }
             }
             Self::VarDecl {
                 id,