+++ crates/icrate/src/additions/Foundation/mod.rs
@@ -32,8 +32,9 @@ pub use self::range::NSRange;
 #[cfg(feature = "Foundation_NSThread")]
 #[cfg(feature = "dispatch")]
 pub use self::thread::MainThreadBound;
+pub use self::thread::MainThreadMarker;
 #[cfg(feature = "Foundation_NSThread")]
-pub use self::thread::{is_main_thread, is_multi_threaded, MainThreadMarker};
+pub use self::thread::{is_main_thread, is_multi_threaded};
 
 #[cfg(feature = "Foundation_NSString")]
 #[doc(inline)]