+++ crates/icrate/src/additions/Foundation/number.rs
@@ -20,11 +20,6 @@ use objc2::encode::Encoding;
 use crate::common::*;
 use crate::Foundation::{CGFloat, NSNumber};
 
-// SAFETY: `NSNumber` is a wrapper around an integer/float/bool, and it is
-// immutable.
-unsafe impl Sync for NSNumber {}
-unsafe impl Send for NSNumber {}
-
 impl UnwindSafe for NSNumber {}
 impl RefUnwindSafe for NSNumber {}